Title: css changes to page
Summary: I edited the my-profile page. The first change I did was getting the my profile column to stick to the top left of the screen. I then fixed the about, contact, and address p tags with hover to give them the desired effect within the css page. In additon, I fixed the about ref landing location on the page, and I also added a pointer to them as well. Lastly, I added it added in edit tags to with a svg icon that changes color.

Test Plan: none

Reviewers: yujie, william

Reviewed By: yujie

Subscribers: william, yujie

Revert Plan: none

Differential Revision: http://13.57.139.206/D635